requirements.txt
Kivy >= 1.11, <1.12
firebase == 3.0.1
python_jwt
gcloud
sseclient
pycryptodome
requests_toolbelt